Test calling a prophecy wihout optional parameter
This commit adds a test to reproduce the reported regression. The test
builds a prophecy for a method which has an optional parameter. The
prophecy doesn't expect that the optional parameter will be called.
Calling this mock without the optional parameter (as expected) will
result in an unexpected call with the optional parameter:

1) phpmock\prophecy\RegressionTest::expectingWithoutOptionalParameter
Prophecy\Exception\Call\UnexpectedCallException: Method call:
  - call("arg1", "optional")
  on Double\phpmock\prophecy\OptionalParameterHolder\P15 was not
  expected, expected calls were:
    - call(exact("arg1"))

See also #1